+++ sdk/cosmos/azure-cosmos-tests/src/test/java/com/azure/cosmos/implementation/RxDocumentClientImplTest.java
@@ -181,6 +181,7 @@ public class RxDocumentClientImplTest {
                 Mockito.any(),
                 Mockito.any(),
                 Mockito.any(),
+                Mockito.any(),
                 Mockito.any()
             ))
             .thenReturn(Flux.just(dummyExecutionContextForQuery(queryResults, headersForQueries)));