+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/rntbd/RntbdClientChannelPool.java
@@ -800,7 +800,7 @@ public final class RntbdClientChannelPool implements ChannelPool {
         // Reduce the connection timeout based on the injected delay
         // The higher delay being injected, then less time left to open connection, then it is easier to get connectionTimeout exception.
         // But we would not want to use a <0 for connectionTimeout setting as netty throw exceptions.
-        long effectiveConnectTimeoutInMillis = Math.max(this.connectTimeoutInMillis - delayInMillis, 10);
+        long effectiveConnectTimeoutInMillis = Math.max(this.connectTimeoutInMillis - delayInMillis, 5);
 
         this.executor.schedule(
             () -> {