+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Release History
 
+### 4.45.2-beta.1 (Unreleased)
+
+#### Bugs Fixed
+* Added capability to mark a region as unavailable when a request is cancelled due to end-to-end timeout and connection issues
+  with the region in the direct connectivity mode. - See [PR 35586](https://github.com/Azure/azure-sdk-for-java/pull/35586)
+
 ### 4.45.1 (2023-05-19)
 
 #### Bugs Fixed